requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.10.0
urllib3==1.26.14
